--- malvis-cnu/src/App.css
@@ -1,4 +1,0 @@
-/* .App {
-  width: 100%;
-  height: 100%;
-} */
