+++ malvis-cnu/src/components/UI/Button.css
@@ -28,7 +28,7 @@
   border-radius: 15px;
   color: white;
   font-size: 1em;
-  padding: 12px 0 12px 0;
+  padding: 12px;
   margin: 12px 0 0 0;
   cursor: pointer;
   width: 100%;
@@ -41,7 +41,31 @@
   border-radius: 15px;
   color: white;
   font-size: 1em;
-  padding: 12px 0 12px 0;
+  padding: 12px;
+  margin: 12px 0 0 0;
+  width: 100%;
+}
+
+.download {
+  background: #ff9f43;
+  border: none;
+  border-radius: 15px;
+  color: white;
+  font-size: 1em;
+  padding: 12px;
+  margin: 12px 0 0 0;
+  width: 100%;
+  cursor: pointer;
+}
+
+.prevent-download {
+  background: #ff9f43;
+  opacity: 0.5;
+  border: none;
+  border-radius: 15px;
+  color: white;
+  font-size: 1em;
+  padding: 12px;
   margin: 12px 0 0 0;
   width: 100%;
 }
