+++ malvis-cnu/src/components/UI/UploadFile.css
@@ -12,7 +12,6 @@
 }
 
 .file-name {
-  margin: 0 16px 0 16px;
   text-align: center;
   overflow: hidden;
   white-space: nowrap;
