+++ malvis-cnu/src/components/aside/Aside.css
@@ -2,6 +2,7 @@
   background: #efefef;
   display: flex;
   flex-direction: column;
+  width: 100%;
   height: 100%;
   padding: 20px 0 0 0;
 }
