+++ malvis-cnu/src/components/aside/Aside.js
@@ -1,15 +1,15 @@
 import "./Aside.css";
 
-const Aside = ({ nodes, results, clusters, clicked, widthStyle }) => {
+const Aside = ({ nodes, results, clusters, clicked }) => {
   return (
     <>
-      <aside className="side-bar" style={widthStyle ? widthStyle : {}}>
+      <aside className="side-bar">
         {results ? (
           <>
             <div style={{ paddingLeft: "8px", height: "5%" }}>
               {"전체 실루엣 계수 : " + results.data.silhouette_score}
             </div>
-            <div style={{ padding: "0 8px 8px 8px" }}>
+            <div style={{ padding: "8px" }}>
               {console.log(results)}
               <div>{"<<< 분석 옵션 >>>"}</div>
               <div>
@@ -34,7 +34,7 @@ const Aside = ({ nodes, results, clusters, clicked, widthStyle }) => {
         ) : null}
         {clusters && clicked === "edge" ? (
           <div
-            style={{ paddingLeft: "8px", height: "62%", overflow: "scroll" }}
+            style={{ paddingLeft: "8px", height: "60%", overflow: "scroll" }}
           >
             <div>{"두 클러스터 간 거리 : " + clusters[2].data.value}</div>
             <div>
