+++ malvis-cnu/src/pages/Layout.css
@@ -1,20 +1,6 @@
 .layout {
+  position: relative;
   display: flex;
   width: 100%;
   height: 100%;
-}
-
-.right-sidebar {
-  background: #efefef;
-  display: flex;
-  flex-direction: column;
-  width: 30%;
-  height: 100%;
-  padding: 20px 40px 0 40px;
-  transition: 1s ease;
-  z-index: 10;
-}
-
-.side-btn {
-  z-index: 10;
 }
